+++ src/analyzer/expr/shape_analyzer.rs
@@ -43,11 +43,11 @@ pub(crate) fn analyze(
         };
 
         if let Some(ref mut current_stmt_offset) = analysis_data.current_stmt_offset {
-            if current_stmt_offset.line != start_pos.line() {
+            if current_stmt_offset.line != start_pos.line() as u32 {
                 *current_stmt_offset = StmtStart {
-                    offset: start_pos.start_offset(),
-                    line: start_pos.line(),
-                    column: start_pos.to_raw_span().start.column() as usize,
+                    offset: start_pos.start_offset() as u32,
+                    line: start_pos.line() as u32,
+                    column: start_pos.to_raw_span().start.column() as u16,
                     add_newline: true,
                 };
             }
