+++ src/analyzer/functionlike_analyzer.rs
@@ -595,10 +595,7 @@ impl<'a> FunctionLikeAnalyzer<'a> {
         if config.remove_fixmes && parent_analysis_data.is_none() {
             for unused_fixme_position in analysis_data.get_unused_hakana_fixme_positions() {
                 analysis_data.add_replacement(
-                    (
-                        unused_fixme_position.0 as usize,
-                        unused_fixme_position.1 as usize,
-                    ),
+                    (unused_fixme_position.0, unused_fixme_position.1),
                     if unused_fixme_position.3 {
                         Replacement::TrimTrailingWhitespace(unused_fixme_position.2)
                     } else {
@@ -1215,9 +1212,9 @@ fn report_unused_expressions(
                         analysis_data.expr_fixme_positions.insert(
                             (pos.start_offset, pos.end_offset),
                             StmtStart {
-                                offset: pos.start_offset as usize,
-                                line: pos.start_line as usize,
-                                column: pos.start_column as usize,
+                                offset: pos.start_offset,
+                                line: pos.start_line,
+                                column: pos.start_column,
                                 add_newline: functionlike_storage.has_multi_line_params(),
                             },
                         );
@@ -1240,7 +1237,7 @@ fn report_unused_expressions(
                             && !config.add_fixmes
                         {
                             if !analysis_data.add_replacement(
-                                (pos.start_offset as usize + 1, pos.start_offset as usize + 1),
+                                (pos.start_offset as u32 + 1, pos.start_offset as u32 + 1),
                                 Replacement::Substitute("_".to_string()),
                             ) {
                                 return;
